Event Details freeze bug fix
There is a reported bug with Ionic that has to do with opening a modal
through a popover.  This causes freezing 100% of the time.  To resolve
this, when the modal is closed via the close button, it reloads the
window which is the calendar.  NOTE: This is a bandaid fix and should be
changed in a future ionic release